pyproject.toml
[project]
name = "llava"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gdcm>=1.1",
    "joblib>=1.5.1",
    "ollama>=0.5.3",
    "pillow>=11.3.0",
    "pydicom>=3.0.1",
    "pylibjpeg>=2.0.1",
    "pylibjpeg-libjpeg>=2.3.0",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
]
